main commit

+++ public/src/components/Logout.jsx
@@ -1,17 +1,21 @@
-import React from 'react';
-import axios from 'axios';
-import { useNavigate } from 'react-router-dom';
-import styled from 'styled-components';
-import { BiPowerOff } from 'react-icons/bi';
-
-function Logout() {
+import React from "react";
+import { useNavigate } from "react-router-dom";
+import { BiPowerOff } from "react-icons/bi";
+import styled from "styled-components";
+import axios from "axios";
+import { logoutRoute } from "../utils/APIRoutes";
+export default function Logout() {
   const navigate = useNavigate();
-
   const handleClick = async () => {
-    localStorage.clear();
-    navigate("/login");
+    const id = await JSON.parse(
+      localStorage.getItem(process.env.REACT_APP_LOCALHOST_KEY)
+    )._id;
+    const data = await axios.get(`${logoutRoute}/${id}`);
+    if (data.status === 200) {
+      localStorage.clear();
+      navigate("/login");
+    }
   };
-
   return (
     <Button onClick={handleClick}>
       <BiPowerOff />
@@ -20,17 +24,16 @@ function Logout() {
 }
 
 const Button = styled.button`
-display: flex;
-justify-content: center;
-align-items: center;
-padding: 0.5rem;
-border-radius: 0.5rem;
-background-color: #9a86f3;
-border: none;
-cursor: pointer;
-svg {
-  font-size: 1.3rem;
-  color: #ebe7ff;
-}`;
-
-export default Logout;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 0.5rem;
+  border-radius: 0.5rem;
+  background-color: #9a86f3;
+  border: none;
+  cursor: pointer;
+  svg {
+    font-size: 1.3rem;
+    color: #ebe7ff;
+  }
+`;
